+++ src/Chart/ChartError.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * League.Period (https://period.thephpleague.com)
+ *
+ * (c) Ignace Nyamagana Butera <nyamsprod@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace League\Period\Chart;
+
+use Throwable;
+
+interface ChartError extends Throwable
+{
+}